HTML tweaks
Consolidate special HTML logic to a mixin. Also avoid using base64 if on pymdownx 6.
